+++ fs/lib/fs.dart
@@ -4,7 +4,7 @@
 /// The fs_shim library.
 ///
 /// This is an awesome library. More dartdocs go here.
-library fs_shim;
+library fs_shim.fs;
 
 import 'dart:async';
 import 'dart:convert';
@@ -29,7 +29,7 @@ abstract class FileSystemEntity {
   /// of these subclasses checks whether the object exists in the file
   /// system object exists and is of the correct type (file, directory,
   /// or link).  To check whether a path points to an object on the
-  /// file system, regardless of the object's type, use the [type]
+  /// file system, regardless of the object's type, use the [FileSystem.type]
   /// static method.
   ///
   Future<bool> exists();
@@ -135,6 +135,9 @@ class FileMode {
 /// snapshotted values returned by the stat() call.
 ///
 abstract class FileStat {
+  /// mode only supported on io file system
+  static const modeNotSupported = -1;
+
   ///
   /// The time of the last change to the data of the file system
   /// object.
@@ -151,6 +154,11 @@ abstract class FileStat {
   /// The size of the file system object.
   ///
   int get size;
+
+  ///
+  /// mode (unix executable only)
+  ///
+  int get mode;
 }
 
 /// Abstract File entity.
@@ -173,19 +181,19 @@ abstract class File extends FileSystemEntity {
   Future<File> create({bool recursive = false});
 
   ///
-  /// Creates a new independent [IOSink] for the file. The
-  /// [IOSink] must be closed when no longer used, to free
+  /// Creates a new independent [StreamSink] for the file. The
+  /// [StreamSink] must be closed when no longer used, to free
   /// system resources.
   ///
-  /// An [IOSink] for a file can be opened in two modes:
+  /// An [StreamSink] for a file can be opened in two modes:
   /// * [FileMode.WRITE]: truncates the file to length zero.
   /// * [FileMode.APPEND]: sets the initial write position to the end
   ///   of the file.
   ///
-  /// When writing strings through the returned [IOSink] the encoding
-  /// specified using [encoding] will be used. The returned [IOSink]
+  /// When writing strings through the returned [StreamSink] the encoding
+  /// specified using [encoding] will be used. The returned [StreamSink]
   /// has an [:encoding:] property which can be changed after the
-  /// [IOSink] has been created.
+  /// [StreamSink] has been created.
   ///
   StreamSink<List<int>> openWrite(
       {FileMode mode = FileMode.write, Encoding encoding = utf8});