+++ fs/lib/fs_shim.dart
@@ -1,1 +1,7 @@
+/// {@canonicalFor fs_io_stub.fileSystemIo}
+library fs_shim;
+
 export 'fs.dart';
+export 'fs_memory.dart' show fileSystemMemory;
+export 'src/io/fs_io.dart' show fileSystemIo;
+export 'src/web/fs_web.dart' show fileSystemWeb;