+++ fs/lib/src/io/fs_io_impl.dart
@@ -1,0 +1,7 @@
+import 'package:fs_shim/fs.dart';
+import 'package:fs_shim/src/io/io_file_system.dart';
+
+FileSystem _fileSystemIo;
+
+/// IO file system.
+FileSystem get fileSystemIo => _fileSystemIo ??= FileSystemIo();