+++ fs/lib/src/io/fs_io_stub.dart
@@ -1,0 +1,8 @@
+import 'package:fs_shim/fs.dart';
+
+/// IO file system.
+FileSystem get fileSystemIo => _stub('fileSystemIo');
+
+T _stub<T>(String message) {
+  throw UnimplementedError(message);
+}