+++ fs/tool/run_browser_test_single.dart
@@ -1,0 +1,13 @@
+import 'package:process_run/shell.dart';
+
+Future main() async {
+  var shell = Shell();
+
+  // pub run build_runner test -- -p chrome test/multiplatform
+  // pub run build_runner test -- -p chrome test/web test/multiplatform
+  await shell.run('''
+
+  pub run build_runner test -- -p chrome test/multiplatform/fs_shim_test.dart
+
+''');
+}